modify: move nomination command to separated class
